chore: refactoring and attempt to make the 2norm more accurate2

+++ src/norms.c
@@ -2,9 +2,9 @@
 #include "math.h"
 
 float euclidean_norm(struct tensor *tensor) {
-	float cum = 0; 
+	long double cum = 0; 
 	for (int i = 0; i < pow(tensor->order,tensor->dimension); i++) {
-		cum += tensor->array[i]*tensor->array[i];
+		cum += ((long double) tensor->array[i])*((long double) tensor->array[i]);
 	}
-	return sqrtf(cum);
+	return (float) sqrtl(cum);
 }
